make popups smaller
